Tweak location of shared function
