+++ packages/playground/wordpress/src/index.ts
@@ -1,11 +1,7 @@
 import { PHP, UniversalPHP } from '@php-wasm/universal';
 import { joinPaths, phpVar } from '@php-wasm/util';
 import { unzipFile } from '@wp-playground/common';
-export {
-	bootWordPress,
-	getFileNotFoundActionForWordPress,
-	looksLikePlaygroundDirectory,
-} from './boot';
+export { bootWordPress, getFileNotFoundActionForWordPress } from './boot';
 export { getLoadedWordPressVersion } from './version-detect';
 
 export * from './version-detect';
@@ -386,6 +382,21 @@ export async function unzipWordPress(php: PHP, wpZip: File) {
 			)
 		);
 	}
+}
+
+/**
+ * Check if the given directory handle directory is a Playground directory.
+ *
+ * @param fileExists Function A function that checks if a file exists relative to an assumed directory.
+ * @returns Promise<boolean> Whether the directory looks like a Playground directory.
+ */
+export async function looksLikePlaygroundDirectory(
+	fileExists: (relativePath: string) => Promise<boolean>
+) {
+	const results = await Promise.all(
+		['wp-config.php', 'wp-content/database/.ht.sqlite'].map(fileExists)
+	);
+	return results.every(Boolean);
 }
 
 function isCleanDirContainingSiteMetadata(path: string, php: PHP) {
